requirements.txt
ipykernel
pandas
numpy
langchain-community
transformers
sentence-transformers
faiss-cpu
langchain
langchain-huggingface
PyPDF2
streamlit
fastapi
requests
huggingface-hub
python-multipart
uvicorn
plotly
python-dotenv